requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
langdetect>=1.0.9
rich>=13.0.0
flask>=3.0.0
gunicorn>=21.0.0
